+++ src/test/java/lotto/domain/ticket/AnalysisTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import lotto.domain.rank.Rank;
+import lotto.domain.winning.Rank;
 
 class AnalysisTest {
