+++ src/test/java/lotto/domain/ticket/TicketsTest.java
@@ -4,12 +4,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 
+import lotto.domain.winning.WinningTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import lotto.AppConfig;
-import lotto.domain.rank.Rank;
+import lotto.domain.winning.Rank;
 import lotto.domain.ticket.generator.CustomTicketGenerator;
 import lotto.dto.TicketDto;
