+++ src/test/java/lotto/domain/ticket/WinningTicketTest.java
@@ -6,12 +6,13 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
 import java.util.stream.Stream;
 
+import lotto.domain.winning.WinningTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import lotto.domain.rank.Rank;
+import lotto.domain.winning.Rank;
 import lotto.exception.LottoException;
 import lotto.exception.LottoExceptionStatus;
